% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{autoreg}
\alias{autoreg}
\alias{autoreg.armmMod}
\title{Extract autoregressive parameters}
\usage{
autoreg(object, ...)

\method{autoreg}{armmMod}(object, se_method = c("quantile", "hpdi"), ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}}

\item{\dots}{Additional arguments, ignored for method compatibility.}

\item{se_method}{A single string, for either using quantiles (\code{"quantile"})
or HPDI (\code{"hpdi"}) to compute the standard errors.
Defaults to \code{"quantile"}.}
}
\value{
AR parameters.
}
\description{
Extract autoregressive parameters
}
\section{Methods (by class)}{
\itemize{
\item \code{autoreg(armmMod)}: Autoregressive parameters for an \code{armmMod} object

}}
