% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{coef.armmMod}
\alias{coef.armmMod}
\title{Coefficients from an \code{armmMod} object}
\usage{
\method{coef}{armmMod}(object, ...)
}
\arguments{
\item{object}{An object of class \code{armmMod}.}

\item{...}{Ignored.}
}
\value{
If random variables present, a list of data frames (1 for each
random-effect grouping variable).
Each data frame contains coefficients for each level of the variable.
If the model doesn't contain random variables, then it returns a
vector of coefficients.
}
\description{
Coefficients from an \code{armmMod} object
}
