% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{errors}
\alias{errors}
\alias{errors.armmMod}
\title{Extract error information}
\usage{
errors(object, ...)

\method{errors}{armmMod}(object, se_method = c("quantile", "hpdi"), ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}}

\item{\dots}{Additional arguments, ignored for method compatibility.}

\item{se_method}{A single string, for either using quantiles (\code{"quantile"})
or HPDI (\code{"hpdi"}) to compute the standard errors.
Defaults to \code{"quantile"}.}
}
\value{
Process and observation error standard deviations.
}
\description{
Extract error information
}
\section{Methods (by class)}{
\itemize{
\item \code{errors(armmMod)}: Process and observation error standard deviations for an \code{armmMod} object

}}
