% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{family.armmMod}
\alias{family.armmMod}
\title{Family for an \code{armmMod} object}
\usage{
\method{family}{armmMod}(object, ...)
}
\arguments{
\item{object}{the function \code{family} accesses the \code{family}
    objects which are stored within objects created by modelling
    functions (e.g., \code{glm}).}

\item{...}{further arguments passed to methods.}
}
\description{
Family for an \code{armmMod} object
}
