% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{fitted.armmMod}
\alias{fitted.armmMod}
\title{Fitted values for an \code{armmMod} object}
\usage{
\method{fitted}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}}

\item{\dots}{Additional arguments, ignored for method compatibility.}
}
\value{
Fitted values.
}
\description{
Fitted values for an \code{armmMod} object
}
