% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_pars}
\alias{generate_pars}
\title{Generate parameter values for simulations.}
\usage{
generate_pars(n_time, n_loc, n_spp, mean_b0, mean_b1, mean_rho, sigma_b0,
  sigma_b1, sigma_rho, sigma_eps, sigma_obs = 0, corr_method = "none")
}
\arguments{
\item{n_time}{Number of time steps.}

\item{n_loc}{Number of locations.}

\item{n_spp}{Number of species.}

\item{mean_b0}{Mean for the b0 parameter relating X to N.}

\item{mean_b1}{Mean for the b1 parameter relating X to N.}

\item{mean_rho}{Mean for the rho parameter relating X to N.
This parameter is on the inverse logit scale.}

\item{sigma_b0}{Standard deviation for the b0 parameter relating X to N.}

\item{sigma_b1}{Standard deviation for the b1 parameter relating X to N.}

\item{sigma_rho}{Standard deviation for the rho parameter relating X to N.
This parameter is on the inverse logit scale.}

\item{sigma_eps}{Standard deviation for the epsilon parameter.}

\item{sigma_obs}{Standard deviation for observation error. Defaults to 0.}

\item{corr_method}{Method for determining correlations between species.
Options are "none", "phylo", or "random". Defaults to "none".}
}
\description{
Generate parameter values to simulate multi-location, multi-species time series data.
}
\examples{
generate_pars(10, 2, 3,
              mean_b0 = log(100),
              mean_b1 = 0.1,
              mean_rho = 0.25,
              sigma_b0 = 0.1,
              sigma_b1 = 0.1,
              sigma_rho = 0.1,
              sigma_eps = 0.1)


}
