% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hpdi}
\alias{hpdi}
\title{Compute Highest Posterior Density Intervals for the parameters in an MCMC sample.}
\usage{
hpdi(input, prob = 0.95)
}
\arguments{
\item{input}{Vector containing the MCMC sample.}

\item{prob}{A number in the interval (0,1) giving the target probability
content of the intervals.
The nominal probability content of the intervals is the multiple of
\code{1/nrow(input)} nearest to \code{prob}.}
}
\description{
This is a C++ version of \code{\link[coda]{HPDinterval}}.
}
