% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\docType{methods}
\name{loo}
\alias{loo}
\alias{loo.armmMod}
\title{Extract LOO estimates from an \code{armmMod} object}
\usage{
\method{loo}{armmMod}(
  x,
  save_psis = FALSE,
  cores = getOption("mc.cores", 1),
  moment_match = FALSE,
  k_threshold = 0.7,
  ...
)
}
\arguments{
\item{x}{A fitted model with class \code{armmMod}.}

\item{save_psis}{Should the intermediate results from \code{\link[loo]{psis}}
  be saved in the returned object? The default is \code{FALSE}. This can be
  useful to avoid repeated computation when using other functions in the
  \pkg{loo} and \pkg{bayesplot} packages.}

\item{cores}{Number of cores to use for parallelization. The default is 1 unless
  \code{cores} is specified or the \code{mc.cores} \link[=options]{option}
  has been set.}

\item{moment_match}{Logical; Whether to use the moment matching algorithm for
  observations with high Pareto k values to improve accuracy. Note:
  because the moment matching algorithm relies on the \code{unconstrain_pars}
  method in RStan it is only available if run in the same R session as fitting the
  model.}

\item{k_threshold}{Threshold value for Pareto k values above which
  the moment matching algorithm is used. If \code{moment_match} is \code{FALSE},
  this is ignored.}

\item{...}{Ignored.}
}
\description{
Extract LOO estimates from an \code{armmMod} object
}
\seealso{
\code{\link[loo]{loo}} \code{\link[rstan]{loo.stanfit}}
}
