% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{model.frame.armmMod}
\alias{model.frame.armmMod}
\title{Extracting the model frame from an \code{armmMod} object}
\usage{
\method{model.frame}{armmMod}(formula, ...)
}
\arguments{
\item{formula}{a model \code{\link[stats]{formula}} or \code{\link[stats]{terms}}
    object or an \R object.}

\item{...}{for \code{model.frame} methods, a mix of further
    arguments such as \code{data}, \code{na.action}, \code{subset} to pass
    to the default method.  Any additional arguments (such as
    \code{offset} and \code{weights} or other named arguments) which
    reach the default method are used to create further columns in the
    model frame, with parenthesised names such as \code{"(offset)"}.

    For \code{get_all_vars}, further named columns to include
    in the model frame.}
}
\description{
Extracting the model frame from an \code{armmMod} object
}
