% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myv_arth}
\alias{myv_arth}
\title{Lake Myvatn arthropods.}
\format{
A data frame with 1120 rows and 6 variables:
\describe{
\item{trans}{}
\item{dist}{}
\item{year}{}
\item{taxon}{}
\item{count}{}
\item{midges}{}
}
}
\source{
\url{https://lter.limnology.wisc.edu/project/ltreb-myvatn-iceland}
}
\usage{
myv_arth
}
\description{
A dataset containing information about Lake Myvatn arthropods.
}
\keyword{datasets}
