% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{nobs.armmMod}
\alias{nobs.armmMod}
\title{Number of observations in an \code{armmMod} object}
\usage{
\method{nobs}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Number of observations in an \code{armmMod} object
}
