% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R, R/mode.R
\name{posterior_mode}
\alias{posterior_mode}
\title{Estimates the marginal parameter mode using kernel density estimation}
\usage{
posterior_mode(x, adjust = 0.1, ...)

posterior_mode(x, adjust = 0.1, ...)
}
\arguments{
\item{x}{numeric vector of posteriors}

\item{adjust}{numeric, passed to density to adjust the bandwidth of the
kernal density}

\item{...}{other parameters to pass to \code{density}}
}
\description{
Estimates the marginal parameter mode using kernel density estimation

Estimates the marginal parameter mode using kernel density estimation
}
