% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\docType{methods}
\name{pp_check}
\alias{pp_check}
\alias{pp_check.armmMod}
\title{Posterior predictive checks for an \code{armmMod} object}
\usage{
\method{pp_check}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{...}{Arguments for function \code{\link[bayesplot]{ppc_dens_overlay}}.}
}
\value{
A \code{ggplot2} object containing the plot.
}
\description{
Posterior predictive checks for an \code{armmMod} object
}
\seealso{
\code{\link[bayesplot]{ppc_dens_overlay}}
}
