% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pops}
\alias{sim_pops}
\title{Simulate populations with competition.}
\usage{
sim_pops(max_t, N0, r, alpha, sigma)
}
\arguments{
\item{max_t}{Number of time points to simulate.}

\item{N0}{Vector of starting population abundances, one for each species.}

\item{r}{Vector of growth rates, one for each species.}

\item{alpha}{Matrix of intra- and inter-specific density dependences.}

\item{sigma}{Standard deviation of process error.}
}
\description{
Simulate populations with competition.
}
\examples{
sim_pops(10, c(10, 10), c(0.5, 0.5), matrix(rep(1e-3, 4), 2, 2), 0.25)

}
