% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pops_ar}
\alias{sim_pops_ar}
\title{Multiple populations simulated using AR1 process.}
\usage{
sim_pops_ar(X, N0_mat, b0_mat, b1_mat, rho_mat, vcv_cube, obs_sigma,
  n_cores = 1)
}
\description{
Input and output N values are logged.
The `X` matrix should have rows associated with a given time point and
columns associated with a given location.
The `vcv_cube` cube should have rows and columns associated with a given species,
and slices associated with a given location.
All other input matrices should have rows associated with a given species and
columns associated with a given location.
`obs_sigma` contains the standard deviations of observation error for each species.
The output array will have rows associated with a given time point,
columns associated with a given species, and 
slices associated with a given location.
}
