% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{summary.armmMod}
\alias{summary.armmMod}
\title{Summary for an \code{armmMod} object.}
\usage{
\method{summary}{armmMod}(
  object,
  se_method = c("quantile", "hpdi"),
  digits = max(3, getOption("digits") - 3),
  ...
)
}
\arguments{
\item{se_method}{A single string, for either using quantiles (\code{"quantile"})
or HPDI (\code{"hpdi"}) to compute the standard errors.
Defaults to \code{"quantile"}.}
}
\description{
Standard errors are based on either quantiles or HPDIs,
and are half the width of the 68\% uncertainty interval.
}
