% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransTrendsPkg-package.R
\docType{package}
\name{TransTrendsPkg-package}
\alias{TransTrendsPkg}
\alias{TransTrendsPkg-package}
\title{TransTrendsPkg: Fit models for trans_trends}
\description{
Some day, more details about what the package does will go here.
}
\author{
\strong{Maintainer}: Lucas A Nell \email{lucas@lucasnell.com}

Authors:
\itemize{
  \item Joseph S Phillips \email{jsphillips2@wisc.edu}
}

}
\keyword{internal}
