% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\docType{methods}
\name{waic}
\alias{waic}
\alias{waic.armmMod}
\title{Extract WAIC from an \code{armmMod} object}
\usage{
\method{waic}{armmMod}(x, ...)
}
\arguments{
\item{x}{A fitted model with class \code{armmMod}.}

\item{\dots}{Ignored.}
}
\description{
Extract WAIC from an \code{armmMod} object
}
\seealso{
\code{\link[loo]{waic}}
}
