% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\docType{methods}
\name{fixef}
\alias{fixef}
\alias{fixef.armmMod}
\alias{fixed.effects}
\title{Extract fixed-effects estimates from an \code{armmMod} object}
\usage{
\method{fixef}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{...}{Ignored.}
}
\value{
A vector of fixed-effects estimates.
}
\description{
Extract fixed-effects estimates from an \code{armmMod} object
}
