% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\docType{methods}
\name{ranef}
\alias{ranef}
\alias{ranef.armmMod}
\alias{random.effects}
\title{Extract random-effects estimates from an \code{armmMod} object}
\usage{
\method{ranef}{armmMod}(object, ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{...}{Ignored.}
}
\value{
A list of random-effects estimates.
}
\description{
Extract random-effects estimates from an \code{armmMod} object
}
