% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armmMod_class.R
\name{residuals.armmMod}
\alias{residuals.armmMod}
\title{Residuals of \code{armmMod} objects}
\usage{
\method{residuals}{armmMod}(object, type = c("response", "deviance", "pearson"), ...)
}
\arguments{
\item{object}{A fitted model with class \code{armmMod}.}

\item{type}{Type of residuals, currently only \code{"response"},
\code{"deviance"}, and \code{"pearson"} are programmed.
Type \code{"deviance"} only works when \code{family(object) == "lnorm_poiss"}.
Defaults to \code{"response"}.}

\item{\dots}{Additional arguments, ignored for method compatibility.}
}
\value{
A vector of residuals.
}
\description{
Getting different types of residuals for \code{armmMod} objects.
}
