% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pops_ar}
\alias{sim_pops_ar}
\title{Multiple populations simulated using AR1 process.}
\usage{
sim_pops_ar(
  X,
  N0_mat,
  b0_mat,
  b1_mat,
  rho_mat,
  vcv_cube,
  obs_sigma,
  n_cores = 1L
)
}
\arguments{
\item{X}{Matrix of environmental variable.
It should have rows associated with a given time point and
columns associated with a given location.}

\item{N0_mat}{Matrix of starting population abundances (\verb{log(# individuals)})
by species and location.}

\item{b0_mat}{Matrix of \eqn{\beta_0} (the population-abundance intercept) values
by species and location.}

\item{b1_mat}{Matrix of \eqn{\beta_1} (the effect of \eqn{X} on \eqn{N}) values
by species and location.}

\item{rho_mat}{Matrix of growth rates by species and location.}

\item{vcv_cube}{Cube representing variance-covariance matrices for process error
among species, one matrix for each location.
It should have rows and columns associated with a given species,
and slices associated with a given location.}

\item{obs_sigma}{Vector of standard deviations of observation error for each species.}

\item{n_cores}{Number of cores to use. Defaults to 1.}
}
\value{
A 3-dimensional array.
The output will have rows associated with a given time point,
columns associated with a given species, and
slices associated with a given location.
}
\description{
Input and output N values are logged.
All input matrices other than \code{X} and \code{vcv_cube} should have rows associated
with a given species and columns associated with a given location.
See descriptions for \code{X} and \code{vcv_cube}.
}
\examples{
X <- matrix(rlnorm(20), 10)
N0 <- matrix(rep(log(10), 6), 3, 2)
b0 <- matrix(rep(log(100), 6), 3, 2)
b1 <- matrix(rep(0.1, 6), 3, 2)
rho <- matrix(rep(0.2, 6), 3, 2)
vcv <- diag(3)
vcv[lower.tri(vcv)] <- vcv[upper.tri(vcv)] <- 0.1
vcv <- array(vcv, dim = c(3, 3, 2))
obs <- rep(0.1, 3)
sim_pops_ar(X, N0, b0, b1, rho, vcv, obs)

}
