% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransTrendsPkg-package.R
\docType{package}
\name{TransTrendsPkg-package}
\alias{TransTrendsPkg}
\alias{TransTrendsPkg-package}
\title{TransTrendsPkg: Fit models for trans_trends}
\description{
This package fits autoregressive, state-space models with random effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucasnell/trans_trends_pkg}
  \item Report bugs at \url{https://github.com/lucasnell/trans_trends_pkg/issues}
}

}
\author{
\strong{Maintainer}: Lucas A Nell \email{lucas@lucasnell.com}

Authors:
\itemize{
  \item Joseph S Phillips \email{jsphillips2@wisc.edu}
}

}
\keyword{internal}
